Move register/unregister to core.controller
